cxgb4: remove cast when saving IPv4 partial checksum (bsc#1074220).
suse-commit: 6fa22ab2f5b00ce138b15244cef28dd6430622a2

+++ drivers/net/ethernet/chelsio/cxgb4/sge.c
@@ -1510,8 +1510,7 @@ out_free:	dev_kfree_skb_any(skb);
 			if (iph->version == 4) {
 				iph->check = 0;
 				iph->tot_len = 0;
-				iph->check = (u16)(~ip_fast_csum((u8 *)iph,
-								 iph->ihl));
+				iph->check = ~ip_fast_csum((u8 *)iph, iph->ihl);
 			}
 			if (skb->ip_summed == CHECKSUM_PARTIAL)
 				cntrl = hwcsum(adap->params.chip, skb);
